+++ src/Recognition/Recognition.jl
@@ -22,9 +22,9 @@ non-negative integer ``k ∈ ℕ``—an optimal permutation that fully minimizes
 ``O(nᵏ)`` time.
 
 The following algorithms are currently supported:
-- Caprara–Salazar-González algorithm ([`CapraraSalazarGonzalez`](@ref))
 - Del Corso–Manzini algorithm ([`DelCorsoManzini`](@ref))
 - Del Corso–Manzini algorithm with perimeter search ([`DelCorsoManziniWithPS`](@ref))
+- Caprara–Salazar-González algorithm ([`CapraraSalazarGonzalez`](@ref))
 - Saxe–Gurari–Sudborough algorithm ([`SaxeGurariSudborough`](@ref))
 - Brute-force search ([`BruteForceSearch`](@ref))
 
@@ -35,11 +35,17 @@ module Recognition
 
 using MatrixBandwidth
 using MatrixBandwidth: NotImplementedError, RectangularMatrixError, StructuralAsymmetryError
-using MatrixBandwidth: connected_components, is_structurally_symmetric, offdiag_nz_support
+using MatrixBandwidth:
+    connected_components,
+    floyd_warshall_shortest_paths,
+    is_structurally_symmetric,
+    offdiag_nz_support
 using MatrixBandwidth: _requires_structural_symmetry, _problem
 
 using Combinatorics: combinations, permutations
 using DataStructures: Queue
+using JuMP
+using HiGHS
 
 export
     # Types
@@ -50,9 +56,9 @@ export
     has_bandwidth_k_ordering,
 
     # Deciders
-    CapraraSalazarGonzalez,
     DelCorsoManzini,
     DelCorsoManziniWithPS,
+    CapraraSalazarGonzalez,
     SaxeGurariSudborough,
     BruteForceSearch
 
@@ -61,12 +67,12 @@ MatrixBandwidth.ALGORITHMS[:Recognition] = []
 include("types.jl")
 include("core.jl")
 
-include("deciders/caprara_salazar_gonzalez.jl")
 # Defines both `DelCorsoManzini` and `DelCorsoManziniWithPS`
 include("deciders/del_corso_manzini.jl")
+include("deciders/caprara_salazar_gonzalez.jl")
 include("deciders/saxe_gurari_sudborough.jl")
 include("deciders/brute_force_search.jl")
 
-const DEFAULT_DECIDER = CapraraSalazarGonzalez()
+const DEFAULT_DECIDER = DelCorsoManzini()
 
 end