+++ test/Recognition/deciders/caprara_salazar_gonzalez.jl
@@ -13,8 +13,50 @@ module TestCapraraSalazarGonzalez
 
 using MatrixBandwidth
 using MatrixBandwidth.Recognition
+using SparseArrays
 using Test
 
-# TODO: Write here
+const MAX_ORDER_LE = 7
+const MAX_ORDER_GT = 8
+const NUM_ITER = 10
+
+@testset "CSG decider – Bandwidth ≤ k (n ≤ $MAX_ORDER_LE)" begin
+    for n in 1:MAX_ORDER_LE, _ in 1:NUM_ITER
+        density = rand()
+        A = sprand(n, n, density)
+        A = A + A' # Ensure structural symmetry
+
+        b = minimize_bandwidth(A, Minimization.BruteForceSearch()).bandwidth
+        k = rand(b:(n - 1))
+
+        res = has_bandwidth_k_ordering(A, k, CapraraSalazarGonzalez())
+        ordering = res.ordering
+
+        @test res.has_ordering
+        @test bandwidth(A[ordering, ordering]) <= k
+    end
+end
+
+@testset "CSG decider – Bandwidth > k (n ≤ $MAX_ORDER_GT)" begin
+    for n in 2:MAX_ORDER_GT, i in 1:NUM_ITER
+        density = rand()
+        A = sprand(n, n, density)
+        A = A + A' # Ensure structural symmetry
+
+        while bandwidth(A) == 0
+            density = rand()
+            A = sprand(n, n, density)
+            A = A + A' # Ensure structural symmetry
+        end
+
+        b = minimize_bandwidth(A, Minimization.BruteForceSearch()).bandwidth
+        k = rand(0:(b - 1))
+
+        res = has_bandwidth_k_ordering(A, k, CapraraSalazarGonzalez())
+
+        @test !res.has_ordering
+        @test isnothing(res.ordering)
+    end
+end
 
 end